1. fix fuckingMake 2. add utility class 3. add parsing for /proc/pid/tasks/*

+++ fuckingMake.sh
@@ -1,2 +1,8 @@
-rm out/*
+if [ -d out ]; then
+	echo cleaning...
+	rm out/*
+else
+	echo clean build
+	mkdir out
+fi
 javac -d out src/*.java
